casting item age to int in memcache api, since it expects all fields to be ints

+++ AppServer/google/appengine/api/memcache/memcache_distributed.py
@@ -285,7 +285,10 @@ class MemcacheService(apiproxy_stub.APIProxyStub):
     stats.set_items(items_total)
     stats.set_bytes(bytes_total)
    
-    stats.set_oldest_item_age(time.time() - time_total / num_servers)
+    # With the Python 2.7 GAE runtime, it expects all fields here to be ints.
+    # Python 2.5 was fine with this being a float, so callers in that runtime
+    # may not be expecting an int.
+    stats.set_oldest_item_age(int(time.time() - time_total / num_servers))
    
   def _GetKey(self, namespace, key):
     """Used to get the Memcache key. It is encoded because the sdk
